+++ rhapsody/multimodal.py
