+++ rhapsody/tests/test_util.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 import pandas as pd
-from minstrel.util import onehot, rank_hits, random_multimodal
+from rhapsody.util import onehot, rank_hits, random_multimodal
 from skbio.util import get_data_path
 import numpy.testing as npt
 import pandas.util.testing as pdt
