+++ rhapsody/util.py
