+++ scripts/rhapsody
@@ -15,16 +15,16 @@ from scipy.sparse import coo_matrix
 import tensorflow as tf
 from tensorflow.contrib.distributions import Multinomial, Normal
 import datetime
-from minstrel.multimodal import Autoencoder, cross_validation
-from minstrel.util import onehot, rank_hits, random_multimodal, split_tables
+from rhapsody.multimodal import Autoencoder, cross_validation
+from rhapsody.util import onehot, rank_hits, random_multimodal, split_tables
 
 
 @click.group()
-def minstrel():
+def rhapsody():
     pass
 
 
-@minstrel.command()
+@rhapsody.command()
 @click.option('--microbe-file',
               help='Input microbial abundances')
 @click.option('--metabolite-file',
@@ -182,7 +182,7 @@ def autoencoder(microbe_file, metabolite_file,
         ranks.to_csv(ranks_file)
 
 
-@minstrel.command()
+@rhapsody.command()
 @click.option('--ranks-file',
               help='Ranks file containing microbe-metabolite rankings')
 @click.option('--k-nearest-neighbors',
@@ -216,4 +216,4 @@ def network(ranks_file, k_nearest_neighbors, node_metadata, edge_metadata, axis)
 
 
 if __name__ == '__main__':
-    minstrel()
+    rhapsody()
